add embedded-db features
