+++ crates/content-base/Cargo.toml
@@ -3,6 +3,10 @@ name = "content-base"
 version = "0.1.0"
 edition = { workspace = true }
 
+[features]
+default = ["embedded-db"]
+embedded-db = ["surrealdb/kv-rocksdb"]
+
 [dependencies]
 tokio = { workspace = true, features = ["full"] }
 async-trait = { workspace = true }
@@ -24,7 +28,7 @@ content-metadata = { path = "../content-metadata" }
 uuid = { workspace = true }
 strum = { workspace = true }
 strum_macros = { workspace = true }
-surrealdb = { version = "2.0.2", features = ["kv-rocksdb"] }
+surrealdb = { version = "2.0.2" }
 futures = { workspace = true }
 futures-util = "0.3.30"
 regex = "1.10.6"
