+++ crates/content-base/src/db/client/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "embedded-db")]
+pub mod embedded;
+
+#[cfg(not(feature = "embedded-db"))]
+pub mod remote;
