+++ crates/content-base/src/db/constant.rs
@@ -1,4 +1,8 @@
 // env
+pub const DATABASE_HOST: &str = "DATABASE_HOST";
+pub const DATABASE_PORT: &str = "DATABASE_PORT";
+pub const DATABASE_USER: &str = "DATABASE_USER";
+pub const DATABASE_PASSWORD: &str = "DATABASE_PASSWORD";
 pub const DATABASE_NS: &str = "DATABASE_NS";
 pub const DATABASE_NAME: &str = "DATABASE_NAME";
